+++ FERN/Views/App Navigation/MainMenuView.swift
@@ -16,6 +16,7 @@ struct MainMenuView: View {
     @State private var map = MapClass()
     @State private var camera = CameraClass()
     @State private var gps = GpsClass()
+    @State private var upload = FileUploadClass()
     
     init() {
         let menuListCoordinator = MenuListBridgingCoordinator()
@@ -34,7 +35,7 @@ struct MainMenuView: View {
                 {
                     // Select Trip Mode (new trip acquisition)
                     NavigationLink {
-                        SelectTripView(map: map, gps: gps, camera: camera)
+                        SelectTripView(map: map, gps: gps, camera: camera, upload: upload)
                             .environment(gps)
                             .navigationTitle("Select Trip Mode")
                     } label: {
@@ -45,7 +46,7 @@ struct MainMenuView: View {
                     }
                     // QC an Uploaded Trip
                     NavigationLink {
-                        SelectMapPlatformView(map: map, gps: gps, camera: camera, mapMode: "View Trip", columnName: "", organismName: "", mapQuery: "query_get_trip_for_apple_map")
+                        SelectMapPlatformView(map: map, gps: gps, camera: camera, upload: upload, mapMode: "View Trip", columnName: "", organismName: "", mapQuery: "query_get_trip_for_apple_map")
                             .environmentObject(menuListBridgingCoordinator)
                             .navigationTitle("Select Platform")
                     } label: {
@@ -56,7 +57,7 @@ struct MainMenuView: View {
                     }
                     // Select a saved route
                     NavigationLink {
-                        ShowListFromDatabaseView(map: map, gps: gps, camera: camera, mapMode: "Traveling Salesman", columnName: "", organismName: "", mapQuery: "query_get_route_for_app", tripType: "")
+                        ShowListFromDatabaseView(map: map, gps: gps, camera: camera, upload: upload, mapMode: "Traveling Salesman", columnName: "", organismName: "", mapQuery: "query_get_route_for_app", tripType: "")
                             .environmentObject(menuListBridgingCoordinator)
                             .navigationTitle("Select Saved Route")
                     } label: {
