+++ FERN/Views/App Navigation/SelectTripTypeView.swift
@@ -15,6 +15,7 @@ struct SelectTripTypeView: View {
     var map: MapClass
     var gps: GpsClass
     var camera: CameraClass
+    var upload: FileUploadClass
     var mapMode: String
     var columnName: String
     var organismName: String
@@ -39,7 +40,7 @@ struct SelectTripTypeView: View {
                 List (self.list) { (item) in
                     NavigationLink(item.name) {
                         // Pass var to view. Query for route does not need a column or organism name.
-                        ShowListFromDatabaseView(map: map, gps: gps, camera: camera, mapMode: mapMode, columnName: columnName, organismName: organismName, mapQuery: mapQuery, tripType: item.name)
+                        ShowListFromDatabaseView(map: map, gps: gps, camera: camera, upload: upload, mapMode: mapMode, columnName: columnName, organismName: organismName, mapQuery: mapQuery, tripType: item.name)
                             .navigationTitle("Apple Map")
                             .environmentObject(menuListBridgingCoordinator)
                     }
